+++ src/components/manifold-resource-card-view/manifold-resource-card-view.spec.ts
@@ -1,3 +1,4 @@
+import { ResourceStatusLabel } from '../../types/graphql';
 import { ManifoldResourceCardView } from './manifold-resource-card-view';
 import { Resource } from '../../spec/mock/marketplace';
 import { Product } from '../../spec/mock/catalog';
@@ -8,7 +9,7 @@ describe('<manifold-resource-card>', () => {
     serviceCard.label = Resource.body.label;
     serviceCard.resourceId = Resource.id;
     serviceCard.logo = Product.body.logo_url;
-    serviceCard.resourceStatus = 'available';
+    serviceCard.resourceStatus = ResourceStatusLabel.Available;
 
     const mock = { emit: jest.fn() };
     serviceCard.resourceClick = mock;
@@ -25,7 +26,7 @@ describe('<manifold-resource-card>', () => {
     serviceCard.label = Resource.body.label;
     serviceCard.resourceId = Resource.id;
     serviceCard.logo = Product.body.logo_url;
-    serviceCard.resourceStatus = 'available';
+    serviceCard.resourceStatus = ResourceStatusLabel.Available;
     serviceCard.resourceLinkFormat = '/resource/:resource';
 
     expect(serviceCard.href).toBe(`/resource/${Resource.body.label}`);