+++ src/components/manifold-resource-list/manifold-resource-list.spec.ts
@@ -2,26 +2,15 @@ import { newSpecPage } from '@stencil/core/testing';
 import fetchMock from 'fetch-mock';
 
 import { ManifoldResourceList } from './manifold-resource-list';
-import { Marketplace } from '../../types/marketplace';
-import { Provisioning } from '../../types/provisioning';
-import { Resource } from '../../spec/mock/marketplace';
 import { connections } from '../../utils/connections';
-import { createRestFetch } from '../../utils/restFetch';
+import { resource } from '../../spec/mock/graphql';
 import { createGraphqlFetch } from '../../utils/graphqlFetch';
 
-const Product = {
-  id: '12345',
-  displayName: 'product',
-  logoUrl: 'https://fillmurray.com/200/200',
-};
-
-const mockGraphqlProducts = {
-  products: {
-    edges: [
-      {
-        node: Product,
-      },
-    ],
+const resources = {
+  data: {
+    resources: {
+      edges: [{ node: resource }, { node: resource }],
+    },
   },
 };
 
@@ -30,11 +19,6 @@ const proto = ManifoldResourceList.prototype as any;
 const oldCallback = proto.componentWillLoad;
 
 proto.componentWillLoad = function() {
-  (this as any).restFetch = createRestFetch({
-    getAuthToken: jest.fn(() => '1234'),
-    wait: () => 10,
-    setAuthToken: jest.fn(),
-  });
   (this as any).graphqlFetch = createGraphqlFetch({
     getAuthToken: jest.fn(() => '1234'),
     wait: () => 10,
@@ -46,28 +30,6 @@ proto.componentWillLoad = function() {
   }
 };
 
-const resources: Marketplace.Resource[] = [
-  {
-    ...Resource,
-    body: {
-      ...Resource.body,
-      product_id: Product.id,
-    },
-  },
-];
-
-const provisionOperation: Provisioning.Operation = {
-  id: '1',
-  body: {
-    resource_id: Resource.id,
-    user_id: '2',
-    type: 'provision',
-    state: 'provision',
-  } as Provisioning.provision,
-  type: 'operation',
-  version: 1,
-};
-
 describe('<manifold-resource-list>', () => {
   window.setInterval = jest.fn();
 
@@ -76,15 +38,12 @@ describe('<manifold-resource-list>', () => {
   });
 
   it('The resources list are rendered if given.', async () => {
-    fetchMock
-      .mock(`${connections.prod.marketplace}/resources/?me`, resources)
-      .mock(connections.prod.graphql, { data: mockGraphqlProducts })
-      .mock(`${connections.prod.provisioning}/operations/?is_deleted=false`, [provisionOperation]);
+    fetchMock.mock(connections.prod.graphql, resources);
 
     const page = await newSpecPage({
       components: [ManifoldResourceList],
       html: `
-          <manifold-resource-list paused=""></manifold-resource-list>
+          <manifold-resource-list paused="true"></manifold-resource-list>
         `,
     });
 
@@ -95,39 +54,38 @@ describe('<manifold-resource-list>', () => {
     // TODO: Add a test that will test that the resources are rendered with their various states
     // to replace the 'processes resources properly' here.
     expect(page.root.shadowRoot.querySelector('.wrapper')).toBeTruthy();
-    expect(page.root.shadowRoot.querySelectorAll('manifold-resource-card-view')).toHaveLength(1);
+    expect(page.root.shadowRoot.querySelectorAll('manifold-resource-card-view')).toHaveLength(
+      resources.data.resources.edges.length
+    );
   });
 
   it('processes resources properly', async () => {
-    fetchMock
-      .mock(`${connections.prod.marketplace}/resources/?me`, resources)
-      .mock(connections.prod.graphql, { data: mockGraphqlProducts })
-      .mock(`${connections.prod.provisioning}/operations/?is_deleted=false`, [provisionOperation]);
+    fetchMock.mock(connections.prod.graphql, resources);
 
     const page = await newSpecPage({
       components: [ManifoldResourceList],
-      html: `<manifold-resource-list paused=""></manifold-resource-list>`,
+      html: `<manifold-resource-list paused="true"></manifold-resource-list>`,
     });
 
     const instance = page.rootInstance as ManifoldResourceList;
 
-    expect(fetchMock.called(`${connections.prod.marketplace}/resources/?me`)).toBe(true);
-    expect(fetchMock.called(connections.prod.graphql)).toBe(true);
-    expect(fetchMock.called(`${connections.prod.provisioning}/operations/?is_deleted=false`)).toBe(
-      true
-    );
+    expect(instance.resources).toHaveLength(resources.data.resources.edges.length);
+    expect(instance.resources).toEqual(resources.data.resources.edges);
+  });
 
-    expect(instance.resources).toHaveLength(1);
-    expect(instance.resources).toEqual([
-      {
-        id: Resource.id,
-        name: Resource.body.name,
-        label: Resource.body.label,
-        logo: Product.logoUrl,
-        logoLabel: Product.displayName,
-        status: 'provision',
-      },
-    ]);
+  it('renders a skeleton when no resources available', async () => {
+    fetchMock.mock(connections.prod.graphql, {});
+
+    const page = await newSpecPage({
+      components: [ManifoldResourceList],
+      html: `<manifold-resource-list paused="true"></manifold-resource-list>`,
+    });
+
+    if (!page.root || !page.root.shadowRoot) {
+      throw new Error('<manifold-resource-list> not found in document');
+    }
+
+    expect(page.root.shadowRoot.querySelectorAll('manifold-resource-card-view')).toHaveLength(4);
   });
 
   describe('v0 API', () => {
@@ -150,15 +108,12 @@ describe('<manifold-resource-list>', () => {
     });
 
     it('slot="loading" is rendered if still loading.', async () => {
-      fetchMock
-        .mock(`${connections.prod.marketplace}/resources/?me`, {})
-        .mock(connections.prod.graphql, { data: [] })
-        .mock(`${connections.prod.provisioning}/operations/?is_deleted=false`, []);
+      fetchMock.mock(connections.prod.graphql, { data: {} });
 
       const page = await newSpecPage({
         components: [ManifoldResourceList],
         html: `
-            <manifold-resource-list paused="">
+            <manifold-resource-list paused="true">
               <div data-test="loading" slot="loading"></div>
             </manifold-resource-list>
           `,
@@ -168,15 +123,12 @@ describe('<manifold-resource-list>', () => {
     });
 
     it('slot="no-resources" is rendered if no resources are found.', async () => {
-      fetchMock
-        .mock(`${connections.prod.marketplace}/resources/?me`, [])
-        .mock(connections.prod.graphql, { data: [] })
-        .mock(`${connections.prod.provisioning}/operations/?is_deleted=false`, []);
+      fetchMock.mock(connections.prod.graphql, { data: resources });
 
       const page = await newSpecPage({
         components: [ManifoldResourceList],
         html: `
-          <manifold-resource-list paused="">
+          <manifold-resource-list paused="true">
             <div data-test="no-resources" slot="no-resources"></div>
           </manifold-resource-list>
         `,